fix: dev_deploy on testnet
adding `test_dev_deploy_testnet` will cost somewhat around 80 Test NEAR
per week for runs in ci,
these are stats based on the number of runs in last week.
And the test needed to be added to detect when some functionality
becomes broken or if it needs some tweaks with additional delays, if
it periodically flakily fails.